+++ project/components/dropdowns/list-card-dropdown.tsx
@@ -3,9 +3,9 @@
 import { ListSchema } from "@/types/lists";
 import { Edit, MoreHorizontal, Trash } from "lucide-react";
 import { Fragment, useState } from "react";
+import { DeleteListModal } from "../modals/delete-list-modal";
+import { UpdateListModal } from "../modals/update-list-modal";
 import Dropdown from "./drop-down";
-import { DeleteListModal } from "./modals/delete-list-modal";
-import { UpdateListModal } from "./modals/update-list-modal";
 
 interface ListCardDropdownProps {
   id: ListSchema["id"];
